+++ src/app/dashboard/templates/page.tsx
@@ -2,7 +2,7 @@
 
 import { useState } from "react";
 import { TemplateCard } from "./TemplateCard";
-import { Header } from "@/components/Header";
+import { Header } from "@/components/blocks/Header";
 import { ToggleGroup, ToggleGroupItem } from "@/components/ui/toggle-group";
 
 import { tagsConfig, initialTemplates } from "./DummyData";
