+++ src/components/blocks/EmptyState.tsx
@@ -1,4 +1,4 @@
-import { Button } from "./ui/button";
+import { Button } from "@/components/ui/button";
 import { Calendar, Database, Plug } from "lucide-react";
 
 interface EmptyStateProps {
