+++ src/components/blocks/InfoCard.tsx
